Move Framebuffer copy utitlity methods into global scope
